Export to IsoplotR

transform into a "fissiontracks" object

Usage

as.fissiontracks(x, spotsize = 40, zeta, ages = FALSE)

Arguments

x

a data.frame of spontaneous and fission track counts and the U-concentration

spotsize

the laser ablation spot size

zeta

the zeta calibration constant extracted from the input data

ages

logical. FALSE if x represents unprocessed data (the default), or TRUE if processed ages

Value

An object of class fissiontracks

Examples

data(standard)
session.zeta <- zeta_ICP(standard, spotsize = 40, mineral = "apatite", standard = "Dur")
data("sample")
as.fissiontracks(sample, spotsize = 40, zeta = session.zeta$zeta)
#> $format
#> [1] 2
#> 
#> $zeta
#>       zeta        std 
#> 0.11831145 0.01182395 
#> 
#> $spotSize
#> [1] 40
#> 
#> $Ns
#>  [1]  7  1  2  1  2  1  3  2  4 NA  2 NA  7  5  5  3 NA  4  1  3  5  2  1  3  6
#> [26]  3  3  6  3  4  2  8  3  2  2  5  3  5  3  6  2  1  2  4  8  3  2  2  1  1
#> [51]  8  4  6  5  3  7  3
#> 
#> $U
#> $U$`1`
#> [1] 0.000905       NA       NA
#> 
#> $U$`2`
#> [1] 0.000828       NA       NA
#> 
#> $U$`3`
#> [1] 0.000774       NA       NA
#> 
#> $U$`4`
#> [1] 0.000416       NA       NA
#> 
#> $U$`5`
#> [1] 0.000724       NA       NA
#> 
#> $U$`6`
#> [1] 0.000122       NA       NA
#> 
#> $U$`7`
#> [1] 0.000929       NA       NA
#> 
#> $U$`8`
#> [1] 0.000521       NA       NA
#> 
#> $U$`9`
#> [1] 0.000907       NA       NA
#> 
#> $U$`10`
#> [1] 5e-04    NA    NA
#> 
#> $U$`11`
#> [1] 0.000585       NA       NA
#> 
#> $U$`12`
#> [1] 0.0157     NA     NA
#> 
#> $U$`13`
#> [1] 0.000819       NA       NA
#> 
#> $U$`14`
#> [1] 0.000944       NA       NA
#> 
#> $U$`15`
#> [1] 0.000704       NA       NA
#> 
#> $U$`16`
#> [1] 0.000904       NA       NA
#> 
#> $U$`17`
#> [1] NA NA NA
#> 
#> $U$`18`
#> [1] 0.000765       NA       NA
#> 
#> $U$`19`
#> [1] 0.000757       NA       NA
#> 
#> $U$`20`
#> [1] 0.000773       NA       NA
#> 
#> $U$`21`
#> [1] 0.000811       NA       NA
#> 
#> $U$`22`
#> [1] 0.002159       NA       NA
#> 
#> $U$`23`
#> [1] 0.00092      NA      NA
#> 
#> $U$`24`
#> [1] 0.000782       NA       NA
#> 
#> $U$`25`
#> [1] 0.000707       NA       NA
#> 
#> $U$`26`
#> [1] 0.000741       NA       NA
#> 
#> $U$`27`
#> [1] 0.000837       NA       NA
#> 
#> $U$`28`
#> [1] 0.000853       NA       NA
#> 
#> $U$`29`
#> [1] 0.00075      NA      NA
#> 
#> $U$`30`
#> [1] 0.000592       NA       NA
#> 
#> $U$`31`
#> [1] 0.001166       NA       NA
#> 
#> $U$`32`
#> [1] 0.001119       NA       NA
#> 
#> $U$`33`
#> [1] 0.00067      NA      NA
#> 
#> $U$`34`
#> [1] 0.000704       NA       NA
#> 
#> $U$`35`
#> [1] 0.000455       NA       NA
#> 
#> $U$`36`
#> [1] 0.000952       NA       NA
#> 
#> $U$`37`
#> [1] 0.000555       NA       NA
#> 
#> $U$`38`
#> [1] 0.001036       NA       NA
#> 
#> $U$`39`
#> [1] 0.000764       NA       NA
#> 
#> $U$`40`
#> [1] 0.001163       NA       NA
#> 
#> $U$`41`
#> [1] 0.000601       NA       NA
#> 
#> $U$`42`
#> [1] 0.000777       NA       NA
#> 
#> $U$`43`
#> [1] 0.000931       NA       NA
#> 
#> $U$`44`
#> [1] 0.000805       NA       NA
#> 
#> $U$`45`
#> [1] 0.000767       NA       NA
#> 
#> $U$`46`
#> [1] 0.000721       NA       NA
#> 
#> $U$`47`
#> [1] 0.00058      NA      NA
#> 
#> $U$`48`
#> [1] 0.000346       NA       NA
#> 
#> $U$`49`
#> [1] 0.000724       NA       NA
#> 
#> $U$`50`
#> [1] 0.000585       NA       NA
#> 
#> $U$`51`
#> [1] 0.000819       NA       NA
#> 
#> $U$`52`
#> [1] 0.000757       NA       NA
#> 
#> $U$`53`
#> [1] 0.000811       NA       NA
#> 
#> $U$`54`
#> [1] 0.000741       NA       NA
#> 
#> $U$`55`
#> [1] 0.000704       NA       NA
#> 
#> $U$`56`
#> [1] 0.001163       NA       NA
#> 
#> $U$`57`
#> [1] 0.00058      NA      NA
#> 
#> 
#> $sU
#> $sU$`1`
#> [1] 1.6e-05      NA      NA
#> 
#> $sU$`2`
#> [1] 1.7e-05      NA      NA
#> 
#> $sU$`3`
#> [1] 1.5e-05      NA      NA
#> 
#> $sU$`4`
#> [1] 2.2e-05      NA      NA
#> 
#> $sU$`5`
#> [1] 1.9e-05      NA      NA
#> 
#> $sU$`6`
#> [1] 1.3e-05      NA      NA
#> 
#> $sU$`7`
#> [1] 3.2e-05      NA      NA
#> 
#> $sU$`8`
#> [1] 2.3e-05      NA      NA
#> 
#> $sU$`9`
#> [1] 2.5e-05      NA      NA
#> 
#> $sU$`10`
#> [1] 0.00019      NA      NA
#> 
#> $sU$`11`
#> [1] 3.9e-05      NA      NA
#> 
#> $sU$`12`
#> [1] 0.0094     NA     NA
#> 
#> $sU$`13`
#> [1] 2e-05    NA    NA
#> 
#> $sU$`14`
#> [1] 2.6e-05      NA      NA
#> 
#> $sU$`15`
#> [1] 1.8e-05      NA      NA
#> 
#> $sU$`16`
#> [1] 2.4e-05      NA      NA
#> 
#> $sU$`17`
#> [1] NA NA NA
#> 
#> $sU$`18`
#> [1] 2.1e-05      NA      NA
#> 
#> $sU$`19`
#> [1] 1.9e-05      NA      NA
#> 
#> $sU$`20`
#> [1] 3.1e-05      NA      NA
#> 
#> $sU$`21`
#> [1] 1.6e-05      NA      NA
#> 
#> $sU$`22`
#> [1] 8.4e-05      NA      NA
#> 
#> $sU$`23`
#> [1] 2.2e-05      NA      NA
#> 
#> $sU$`24`
#> [1] 1.8e-05      NA      NA
#> 
#> $sU$`25`
#> [1] 2e-05    NA    NA
#> 
#> $sU$`26`
#> [1] 3.1e-05      NA      NA
#> 
#> $sU$`27`
#> [1] 2.7e-05      NA      NA
#> 
#> $sU$`28`
#> [1] 1.6e-05      NA      NA
#> 
#> $sU$`29`
#> [1] 1.9e-05      NA      NA
#> 
#> $sU$`30`
#> [1] 3.3e-05      NA      NA
#> 
#> $sU$`31`
#> [1] 3.1e-05      NA      NA
#> 
#> $sU$`32`
#> [1] 3.2e-05      NA      NA
#> 
#> $sU$`33`
#> [1] 2e-05    NA    NA
#> 
#> $sU$`34`
#> [1] 1.3e-05      NA      NA
#> 
#> $sU$`35`
#> [1] 2.1e-05      NA      NA
#> 
#> $sU$`36`
#> [1] 1.3e-05      NA      NA
#> 
#> $sU$`37`
#> [1] 1.5e-05      NA      NA
#> 
#> $sU$`38`
#> [1] 2.6e-05      NA      NA
#> 
#> $sU$`39`
#> [1] 1.6e-05      NA      NA
#> 
#> $sU$`40`
#> [1] 2.2e-05      NA      NA
#> 
#> $sU$`41`
#> [1] 1.7e-05      NA      NA
#> 
#> $sU$`42`
#> [1] 1.8e-05      NA      NA
#> 
#> $sU$`43`
#> [1] 2e-05    NA    NA
#> 
#> $sU$`44`
#> [1] 3.9e-05      NA      NA
#> 
#> $sU$`45`
#> [1] 6.4e-05      NA      NA
#> 
#> $sU$`46`
#> [1] 2.5e-05      NA      NA
#> 
#> $sU$`47`
#> [1] 1.8e-05      NA      NA
#> 
#> $sU$`48`
#> [1] 1.7e-05      NA      NA
#> 
#> $sU$`49`
#> [1] 1.9e-05      NA      NA
#> 
#> $sU$`50`
#> [1] 3.9e-05      NA      NA
#> 
#> $sU$`51`
#> [1] 2e-05    NA    NA
#> 
#> $sU$`52`
#> [1] 1.9e-05      NA      NA
#> 
#> $sU$`53`
#> [1] 1.6e-05      NA      NA
#> 
#> $sU$`54`
#> [1] 3.1e-05      NA      NA
#> 
#> $sU$`55`
#> [1] 1.3e-05      NA      NA
#> 
#> $sU$`56`
#> [1] 2.2e-05      NA      NA
#> 
#> $sU$`57`
#> [1] 1.8e-05      NA      NA
#> 
#> 
#> attr(,"class")
#> [1] "fissiontracks"