Zeta calibration coefficient for LA-ICP-MS fission track dating

Calibrate the ICP-session zeta factor using the number of spontaneous tracks, the area over which these were counted and one ore more U/Ca, U/Si or U-concentration measurements (in ppm) and their analytical uncertainties.

Usage

zeta_ICP(
  x,
  spotsize = 40,
  mineral = c("zircon", "apatite"),
  standard = c("Dur", "FCT"),
  tst,
  lambda = pkg.env$lambda[1],
  g = pkg.env$g,
  ...
)

Arguments

x

data.frame containing counts of spontaneous tracks, U concentrations, and the uncertainties of U concentrations

spotsize

the laser ablation spot size (diameter in micrometer)

mineral

Mineral of the standard, one of 'zircon' or 'apatite'

standard

Name of the standard, one of 'Dur' or 'FCT'

tst

(optional) two-column vector giving the true age and its standard error. overrides mineral and standard

lambda

(optional) decay constant

g

(optional) geometry factor (default: 0.5)

...

additional arguments

Value

list containing: two-element vector with the session zeta and its standard error

References

Hasebe, N., Carter, A., Hurford, A.J., Arai, S., 2009. The effect of chemical etching on LA-ICP-MS analysis in determining uranium concentration for fission-track chronometry. Geol. Soc. Lond. Spec. Publ. 324 (1), 37–46. doi:10.1144/SP324.3

Examples

data("standard")
zeta_ICP(standard, spotsize = 40, mineral = "apatite", standard = "Dur")
#> $data
#> # A tibble: 34 × 6
#>    Grain     Ns UO           U       sU comments
#>    <chr>  <dbl> <lgl>    <dbl>    <dbl> <lgl>   
#>  1 DUR-1      4 NA    0.000496 0.000016 NA      
#>  2 DUR-2      3 NA    0.00052  0.000018 NA      
#>  3 DUR-3      1 NA    0.000476 0.000014 NA      
#>  4 DUR-4      3 NA    0.000465 0.000014 NA      
#>  5 DUR-5      6 NA    0.000471 0.000012 NA      
#>  6 DUR-6      3 NA    0.000469 0.000014 NA      
#>  7 DUR-7      4 NA    0.000471 0.000013 NA      
#>  8 DUR-8      2 NA    0.000495 0.000016 NA      
#>  9 DUR-9      4 NA    0.000477 0.000018 NA      
#> 10 DUR-10     2 NA    0.000461 0.000015 NA      
#> # ℹ 24 more rows
#> 
#> $results
#>         rhos       zeta
#> 1  318309.89 0.09822082
#> 2  238732.41 0.13729792
#> 3   79577.47 0.37704122
#> 4  238732.41 0.12277603
#> 5  477464.83 0.06218012
#> 6  238732.41 0.12383217
#> 7  318309.89 0.09327018
#> 8  159154.94 0.19604559
#> 9  318309.89 0.09445833
#> 10 159154.94 0.18257984
#> 11 238732.41 0.12013568
#> 12 159154.94 0.17069828
#> 13 159154.94 0.17307460
#> 14 238732.41 0.11960761
#> 15 397887.36 0.07176457
#> 16 238732.41 0.11142255
#> 17 557042.30 0.04933673
#> 18 238732.41 0.12198393
#> 19 318309.89 0.09406228
#> 20 238732.41 0.12673655
#> 21 238732.41 0.17188644
#> 22  79577.47 0.48951991
#> 23  79577.47 0.47130153
#> 24 238732.41 0.15683648
#> 25 318309.89 0.09267610
#> 26  79577.47 0.36436757
#> 27 397887.36 0.07667561
#> 28 477464.83 0.05835162
#> 29 318309.89 0.09564649
#> 30 238732.41 0.12832075
#> 31 238732.41 0.12383217
#> 32 238732.41 0.12515234
#> 33 159154.94 0.18178773
#> 34 318309.89 0.08752743
#> 
#> $zeta
#>       zeta        std 
#> 0.11831145 0.01182395 
#>